requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
behave==1.2.6
behave-django==1.3.0
cfgv==3.3.0
distlib==0.3.2
Django==3.1.7
django-cors-headers==3.2.1
djangorestframework==3.12.4
entrypoints==0.3
environs==7.1.0
filelock==3.0.12
flake8==3.7.9
flake8-docstrings==1.5.0
identify==2.2.11
importlib-metadata==4.6.1
importlib-resources==5.2.0
jsonschema==3.2.0
marshmallow==3.12.2
mccabe==0.6.1
nodeenv==1.6.0
parse==1.19.0
parse-type==0.5.2
platformdirs==2.0.2
pre-commit==2.13.0
psycopg2==2.8.6
pycodestyle==2.5.0
pydocstyle==6.1.1
pyflakes==2.1.1
pyrsistent==0.18.0
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.6.0
zipp==3.5.0
